backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
pillow
opencv-python-headless
numpy
mediapipe
